requirements.txt
aiohttp==3.8.4
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.1.0
discord-py-interactions==4.3.0
discord-py-slash-command==4.2.1
discord.py==2.2.2
frozenlist==1.3.1
idna==3.3
multidict==6.0.2
psutil==5.9.1
pycparser==2.21
PyNaCl==1.4.0
python-dotenv==0.20.0
six==1.16.0
typing_extensions==4.1.1
yarl==1.7.2